GRADLE-2111: Renamed 'groups' to 'categories'. Added integration test. Made CategoryFilter top-level class.
