+++ subprojects/plugins/src/integTest/resources/org/gradle/testing/junit/JUnitIntegrationTest/supportsTestCategories/build.gradle
@@ -1,0 +1,16 @@
+apply plugin: "java"
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testCompile "junit:junit:4.11"
+}
+
+test {
+    useJUnit {
+        includeCategories 'org.gradle.CategoryA'
+        excludeCategories 'org.gradle.CategoryC'
+    }
+}
