+++ subprojects/plugins/src/integTest/resources/org/gradle/testing/junit/JUnitIntegrationTest/supportsTestCategories/src/test/java/org/gradle/CategoryA.java
@@ -1,0 +1,4 @@
+package org.gradle;
+
+public interface CategoryA {
+}
